refactor: Add EmployeeQualification class to reduce complexity of Employee class and do minor variable name changes
